+++ backend/core/src/auth/services/authz.service.ts
@@ -3,6 +3,7 @@ import { newEnforcer } from "casbin"
 import path from "path"
 import { User } from "../entities/user.entity"
 import { Listing } from "../../listings/entities/listing.entity"
+import { UserRoleEnum } from "../enum/user-role-enum"
 
 export enum authzActions {
   create = "create",
@@ -39,26 +40,35 @@ export class AuthzService {
 
     // Get User roles and add them to our enforcer
     if (user) {
-      await Promise.all(user.roles.map((r) => e.addRoleForUser(user.id, r)))
-    }
+      if (user.roles?.isAdmin) {
+        await e.addRoleForUser(user.id, UserRoleEnum.admin)
+      }
+      if (user.roles?.isPartner) {
+        await e.addRoleForUser(user.id, UserRoleEnum.partner)
+      }
+      await e.addRoleForUser(user.id, UserRoleEnum.user)
 
-    if (user) {
       // NOTE This normally should be in authz_policy.csv, but casbin does not support expressions on arrays.
       //  Permissions for a leasing agent on applications are there defined here programatically.
       //  A User becomes a leasing agent for a given listing if he has a relation (M:N) with it.
       //  User side this is expressed by 'leasingAgentInListings' property.
       await Promise.all(
-        user?.leasingAgentInListings.map((listing: Listing) =>
-          e.addPermissionForUser(
+        user?.leasingAgentInListings.map((listing: Listing) => {
+          void e.addPermissionForUser(
             user.id,
             "application",
             `!r.obj || r.obj.listing_id == '${listing.id}'`,
             `(${authzActions.read}|${authzActions.create}|${authzActions.update}|${authzActions.delete})`
           )
-        )
+          void e.addPermissionForUser(
+            user.id,
+            "listing",
+            `!r.obj || r.obj.listing_id == '${listing.id}'`,
+            `(${authzActions.read}|${authzActions.update})`
+          )
+        })
       )
     }
-
     return e.enforce(user ? user.id : "anonymous", type, action, obj)
   }
 
@@ -72,7 +82,7 @@ export class AuthzService {
    * ABAC logic. Note that a limitation in casbin seems to only allows for property retrieval one level deep on this
    * object (e.g. obj.prop.value wouldn't work).
    */
-  public async canOrThrow(user: User | undefined, type: string, action: string, obj?: any) {
+  public async canOrThrow(user: User | undefined, type: string, action: string, obj?: unknown) {
     if (!(await this.can(user, type, action, obj))) {
       throw new HttpException("Forbidden", HttpStatus.FORBIDDEN)
     }