+++ backend/core/src/listings/entities/listing.entity.ts
@@ -383,9 +383,6 @@ class Listing extends BaseEntity {
   @Expose()
   applicationConfig?: Record<string, unknown>
 
-  @Expose()
-  applicationCount?: number
-
   @Column({ type: "boolean" })
   @Expose()
   @IsBoolean({ groups: [ValidationsGroupsEnum.default] })
@@ -496,6 +493,12 @@ class Listing extends BaseEntity {
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
   @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
   amiPercentageMax?: number | null
+
+  @Column({ type: "boolean", nullable: true })
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsBoolean({ groups: [ValidationsGroupsEnum.default] })
+  customMapPin?: boolean | null
 }
 
 export { Listing as default, Listing }