+++ backend/core/src/seed.ts
@@ -17,21 +17,17 @@ import { ListingDefaultOpenSoonSeed } from "./seeds/listings/listing-default-ope
 import { ListingDefaultOnePreferenceSeed } from "./seeds/listings/listing-default-one-preference-seed"
 import { ListingDefaultNoPreferenceSeed } from "./seeds/listings/listing-default-no-preference-seed"
 import { ListingTritonSeed } from "./seeds/listings/listing-triton-seed"
-import { ListingDefaultBmrChartSeed } from "./seeds/listings/listing-default-bmr-chart-seed"
 import { ApplicationMethodsService } from "./application-methods/application-methods.service"
 import { ApplicationMethodType } from "./application-methods/types/application-method-type-enum"
-import { PaperApplicationsService } from "./paper-applications/paper-applications.service"
-import { Language } from "./shared/types/language-enum"
-import { AssetsService } from "./assets/services/assets.service"
 import { AuthContext } from "./auth/types/auth-context"
 import { ListingDefaultReservedSeed } from "./seeds/listings/listing-default-reserved-seed"
-import { ListingDefaultFCFSSeed } from "./seeds/listings/listing-default-fcfs-seed"
 import { Listing10158Seed } from "./seeds/listings/listing-detroit-10158"
 import { Listing10157Seed } from "./seeds/listings/listing-detroit-10157"
 import { Listing10147Seed } from "./seeds/listings/listing-detroit-10147"
 import { Listing10145Seed } from "./seeds/listings/listing-detroit-10145"
 import { CountyCode } from "./shared/types/county-code"
 import { ListingTreymoreSeed } from "./seeds/listings/listing-detroit-treymore"
+import { UserRoles } from "./auth/entities/user-roles.entity"
 
 const argv = yargs.scriptName("seed").options({
   test: { type: "boolean", default: false },
@@ -59,7 +55,10 @@ export function getSeedListingsCount() {
   return listingSeeds.length
 }
 
-export async function createLeasingAgents(app: INestApplicationContext) {
+export async function createLeasingAgents(
+  app: INestApplicationContext,
+  rolesRepo: Repository<UserRoles>
+) {
   const usersService = await app.resolve<UserService>(UserService)
   const leasingAgents = await Promise.all(
     defaultLeasingAgents.map(
@@ -68,53 +67,23 @@ export async function createLeasingAgents(app: INestApplicationContext) {
   )
   await Promise.all([
     leasingAgents.map(async (agent: User) => {
+      const roles: UserRoles = { user: agent, isPartner: true }
+      await rolesRepo.save(roles)
       await usersService.confirm({ token: agent.confirmationToken })
     }),
   ])
   return leasingAgents
 }
 
-async function createApplicationMethods(app: INestApplicationContext) {
-  const assetsService = await app.resolve<AssetsService>(AssetsService)
-  const englishFileAsset = await assetsService.create({
-    fileId: "englishFileId",
-    label: "English paper application",
-  })
-  const paperApplicationsService = await app.resolve<PaperApplicationsService>(
-    PaperApplicationsService
-  )
-  const englishPaperApplication = await paperApplicationsService.create({
-    language: Language.en,
-    file: englishFileAsset,
-  })
-  const applicationMethodsService = await app.resolve<ApplicationMethodsService>(
-    ApplicationMethodsService
-  )
-
-  await applicationMethodsService.create({
-    type: ApplicationMethodType.FileDownload,
-    acceptsPostmarkedApplications: false,
-    externalReference: "https://bit.ly/2wH6dLF",
-    label: "English",
-    paperApplications: [englishPaperApplication],
-  })
-
-  await applicationMethodsService.create({
-    type: ApplicationMethodType.Internal,
-    acceptsPostmarkedApplications: false,
-    externalReference: "",
-    label: "Label",
-    paperApplications: [],
-  })
-}
-
-const seedListings = async (app: INestApplicationContext) => {
+const seedListings = async (app: INestApplicationContext, rolesRepo: Repository<UserRoles>) => {
   const seeds = []
-  const leasingAgents = await createLeasingAgents(app)
-  await createApplicationMethods(app)
+  const leasingAgents = await createLeasingAgents(app, rolesRepo)
 
   const allSeeds = listingSeeds.map((listingSeed) => app.get<ListingDefaultSeed>(listingSeed))
   const listingRepository = app.get<Repository<Listing>>(getRepositoryToken(Listing))
+  const applicationMethodsService = await app.resolve<ApplicationMethodsService>(
+    ApplicationMethodsService
+  )
 
   for (const [index, listingSeed] of allSeeds.entries()) {
     const listing = await listingSeed.seed()
@@ -122,8 +91,18 @@ const seedListings = async (app: INestApplicationContext) => {
     if (listing.countyCode !== CountyCode.detroit) {
       const everyOtherAgent = index % 2 ? leasingAgents[0] : leasingAgents[1]
       listing.leasingAgents = [everyOtherAgent]
+      const applicationMethods = await applicationMethodsService.create({
+        type: ApplicationMethodType.Internal,
+        acceptsPostmarkedApplications: false,
+        externalReference: "",
+        label: "Label",
+        paperApplications: [],
+        listing: listing,
+      })
+      listing.applicationMethods = [applicationMethods]
       await listingRepository.save(listing)
     }
+
     seeds.push(listing)
   }
 
@@ -132,10 +111,13 @@ const seedListings = async (app: INestApplicationContext) => {
 
 async function seed() {
   const app = await NestFactory.create(SeederModule.forRoot({ test: argv.test }))
+  // Starts listening for shutdown hooks
+  app.enableShutdownHooks()
   const userService = await app.resolve<UserService>(UserService)
 
   const userRepo = app.get<Repository<User>>(getRepositoryToken(User))
-  const listings = await seedListings(app)
+  const rolesRepo = app.get<Repository<UserRoles>>(getRepositoryToken(UserRoles))
+  const listings = await seedListings(app, rolesRepo)
 
   const user1 = await userService.createUser(
     plainToClass(UserCreateDto, {
@@ -190,8 +172,10 @@ async function seed() {
     }
   }
 
-  admin.isAdmin = true
   await userRepo.save(admin)
+  const roles: UserRoles = { user: admin, isPartner: true, isAdmin: true }
+  await rolesRepo.save(roles)
+
   await userService.confirm({ token: admin.confirmationToken })
   await app.close()
 }