+++ backend/core/test/authz/authz.e2e-spec.ts
@@ -101,8 +101,11 @@ describe("Authz", () => {
       await supertest(app.getHttpServer()).get(userEndpoint).expect(403)
     })
 
-    it(`should not allow anonymous user to PUT user`, async () => {
-      await supertest(app.getHttpServer()).get(userEndpoint).expect(403)
+    it(`should allow a logged in user to GET to get any user profile`, async () => {
+      await supertest(app.getHttpServer())
+        .get(userEndpoint)
+        .set(...setAuthorization(userAccessToken))
+        .expect(200)
     })
 
     it(`should allow anonymous user to CREATE a user`, async () => {
@@ -113,6 +116,12 @@ describe("Authz", () => {
   describe("applications", () => {
     it("should not allow anonymous user to GET applications", async () => {
       await supertest(app.getHttpServer()).get(applicationsEndpoint).expect(403)
+    })
+    it("should allow logged in user to GET applications", async () => {
+      await supertest(app.getHttpServer())
+        .get(applicationsEndpoint)
+        .set(...setAuthorization(userAccessToken))
+        .expect(200)
     })
     it("should not allow anonymous user to GET CSV applications", async () => {
       await supertest(app.getHttpServer())