+++ backend/core/types/src/backend-swagger.ts
@@ -3411,10 +3411,18 @@ export interface LoginResponse {
   accessToken: string
 }
 
-export interface User {
+export interface UserRoles {
   /**  */
-  roles: UserRole[]
+  user: User
 
+  /**  */
+  isAdmin: boolean
+
+  /**  */
+  isPartner: boolean
+}
+
+export interface User {
   /**  */
   language?: Language
 
@@ -3447,6 +3455,9 @@ export interface User {
 
   /**  */
   updatedAt: Date
+
+  /**  */
+  roles?: CombinedUserRolesTypes
 }
 
 export interface UserCreate {
@@ -4093,6 +4104,9 @@ export interface Listing {
   buildingTotalUnits?: number
 
   /**  */
+  customMapPin?: boolean
+
+  /**  */
   developer?: string
 
   /**  */
@@ -4232,9 +4246,6 @@ export interface Listing {
 
   /**  */
   applicationConfig?: object
-
-  /**  */
-  applicationCount?: number
 
   /**  */
   displayWaitlistSize: boolean
@@ -4443,6 +4454,9 @@ export interface ListingCreate {
 
   /**  */
   buildingTotalUnits?: number
+
+  /**  */
+  customMapPin?: boolean
 
   /**  */
   developer?: string
@@ -4822,6 +4836,9 @@ export interface ListingUpdate {
 
   /**  */
   buildingTotalUnits?: number
+
+  /**  */
+  customMapPin?: boolean
 
   /**  */
   developer?: string
@@ -5403,10 +5420,6 @@ export enum EnumApplicationsApiExtraModelOrder {
   "ASC" = "ASC",
   "DESC" = "DESC",
 }
-export enum UserRole {
-  "user" = "user",
-  "admin" = "admin",
-}
 export enum EnumListingFilterParamsComparison {
   "=" = "=",
   "<>" = "<>",
@@ -5468,5 +5481,6 @@ export type CombinedApplicationMailingAddressTypes = AddressUpdate
 export type CombinedImageTypes = AssetCreate
 export type CombinedLeasingAgentAddressTypes = AddressUpdate
 export type CombinedResultTypes = AssetCreate
+export type CombinedUserRolesTypes = UserRoles
 export type CombinedWhatToExpectTypes = WhatToExpect
 export type CombinedJurisdictionTypes = Id