+++ sites/partners/pages/index.tsx
@@ -9,7 +9,7 @@ import {
   LocalizedLink,
 } from "@bloom-housing/ui-components"
 import moment from "moment"
-import { UserRole, Listing } from "@bloom-housing/backend-core/types"
+import { Listing } from "@bloom-housing/backend-core/types"
 import { AgGridReact } from "ag-grid-react"
 import { GridOptions } from "ag-grid-community"
 
@@ -21,7 +21,7 @@ import { Router, useRouter } from "next/router"
 export default function ListingsList() {
   const { profile } = useContext(AuthContext)
   const leasingAgentInListings = profile.leasingAgentInListings?.map((item) => item.id)
-  const isAdmin = profile.roles.includes(UserRole.admin)
+  const isAdmin = profile.roles?.isAdmin || false
   class formatLinkCell {
     link: HTMLAnchorElement
 
@@ -81,11 +81,21 @@ export default function ListingsList() {
   const columnDefs = useMemo(() => {
     const columns = [
       {
-        headerName: t("listings.applications"),
-        field: isAdmin ? "applicationCount" : "name",
+        headerName: t("listings.listingName"),
+        field: "name",
         sortable: false,
         filter: false,
         resizable: true,
+        cellRenderer: "ListingsLink",
+      },
+      {
+        headerName: t("listings.listingStatusText"),
+        field: "status",
+        sortable: false,
+        filter: false,
+        resizable: true,
+        flex: 1,
+        valueFormatter: ({ value }) => t(`listings.${value}`),
         cellRenderer: "ApplicationsLink",
       },
       {
@@ -118,34 +128,15 @@ export default function ListingsList() {
         resizable: true,
         cellRenderer: "formatWaitlistStatus",
       },
-      {
-        headerName: t("listings.listingStatusText"),
-        field: "status",
-        sortable: false,
-        filter: false,
-        resizable: true,
-        flex: 1,
-        valueFormatter: ({ value }) => t(`listings.${value}`),
-      },
     ]
-    if (isAdmin) {
-      columns.unshift({
-        headerName: t("listings.listingName"),
-        field: "name",
-        sortable: false,
-        filter: false,
-        resizable: true,
-        cellRenderer: "ListingsLink",
-      })
-    }
     return columns
-  }, [isAdmin])
+  }, [])
 
   const { listingDtos, listingsLoading, listingsError } = useListingsData()
 
   // filter listings to show items depends on user role
   const filteredListings = useMemo(() => {
-    if (profile.roles.includes(UserRole.admin)) return listingDtos
+    if (isAdmin) return listingDtos
 
     return listingDtos?.reduce((acc, curr) => {
       if (leasingAgentInListings.includes(curr.id)) {
@@ -154,7 +145,7 @@ export default function ListingsList() {
 
       return acc
     }, []) as Listing[]
-  }, [leasingAgentInListings, listingDtos, profile.roles])
+  }, [leasingAgentInListings, listingDtos, isAdmin])
 
   if (listingsLoading) return "Loading..."
   if (listingsError) return "An error has occurred."